backend/requirements.txt
flask==3.0.3
mysql-connector-python==9.0.0
flask-cors==4.0.0
python-dotenv==1.0.1
PyJWT==2.9.0
